Remove autoignition-tpe, add mixture-tpe.
